taimen: Add Google platform certificate
These apps are normally present in vendor/app, but they carry the Google
platform certificate, so we had to copy them to the system image we build
so they would be re-signed with our keys.

Now we can declare the Google key as auxiliary to ours and allow apps with
the hardcoded signature to pass.

Change-Id: Icab3fa585d2c186f318999b98dfe3b1eedd3fd1b
Signed-off-by: Joe Maples <joe@frap129.org>